+++ pprl-scaling-framework-mapreduce/pprl-scaling-framework-mapreduce-blocking/src/main/java/gr/upatras/ceid/pprl/mapreduce/HammingLSHFPSToolV2.java
@@ -98,7 +98,7 @@ public class HammingLSHFPSToolV2 extends Configured implements Tool {
         conf.set(CommonKeys.BOB_UID,bobUidFieldName);
         conf.setInt(CommonKeys.BLOCKING_GROUP_COUNT, L);
         conf.setStrings(CommonKeys.BLOCKING_KEYS,blocking.groupsAsStrings());
-        conf.setDouble(CommonKeys.HAMMING_THRESHOLD,hammingThrehold);
+        conf.setInt(CommonKeys.HAMMING_THRESHOLD,hammingThrehold);
         conf.setInt(CommonKeys.FREQUENT_PAIR_LIMIT, C);
 
         conf.setInt("mapreduce.map.memory.mb", 1024);
